fungující dialog včetně ResponseType
