+++ FileCommander/GUI/InputDialogWindow.cs
@@ -4,32 +4,43 @@ using Gtk;
 
 public class InputDialogWindow : Dialog
 {
-    public static string _path { get; set; }
+    private static string _path { get; set; }
 
     private static Dialog _dialog { get; set; }
 
     public InputDialogWindow(string title)
     {
-        _dialog = new Gtk.Dialog(title, this, DialogFlags.DestroyWithParent, Stock.Cancel, ButtonsType.Cancel, Stock.Ok,
+        _dialog = new Dialog(title, this, DialogFlags.DestroyWithParent, Stock.Cancel, ButtonsType.Cancel, Stock.Ok,
             ButtonsType.Ok);
 
         var box = new VBox();
         _dialog.Add(box);
 
         var entry = new Entry();
-        entry.Changed += OnEntry!;
+        //TODO přidat validaci vstupu
+        //TODO nechat jako delegate, nebo změnit na lambdu?
+        _dialog.Response += delegate(object o, ResponseArgs args)
+        {
+            Console.WriteLine(args.ResponseId.ToString());
+            if ((int) args.ResponseId == 1) //OK
+            {
+                _path = entry.Text;
+            }
+            else
+            {
+                NullPath();
+            }
+        };
         _dialog.ContentArea.PackStart(entry, true, true, 0);
         _dialog.ShowAll();
         _dialog.Run();
         _dialog.Destroy();
-
     }
 
-    public Dialog GetDialog() => _dialog;
+    public static string GetPath() => _path;
 
-    void OnEntry(object sender, EventArgs args)
+    public static void NullPath()
     {
-        Entry entry = (Entry) sender;
-        _path = entry.Text;
+        _path = "";
     }
 }
