+++ FileCommander/GUI/UserInterface.cs
@@ -1,89 +1,298 @@
 namespace FileCommander.GUI;
 
-using Gtk;
+using static FunctionController;
 using System;
+using System.IO;
+using Gtk;
 
-class SharpApp : Window
+public class IconApp : Window
 {
-    public SharpApp() : base("File Commander")
+    const int ColPath = 0;
+    const int ColDisplayName = 1;
+    const int ColPixbuf = 2;
+    const int ColIsDirectory = 3;
+
+
+    public static DirectoryInfo LeftRoot = new(Environment.GetFolderPath(Environment.SpecialFolder.Personal)),
+        RightRoot = new(Environment.GetFolderPath(Environment.SpecialFolder.Personal));
+
+    private static readonly Gdk.Pixbuf FileIcon = GetIcon(Stock.File), DirIcon = GetIcon(Stock.Open);
+    public static ListStore LeftStore, RightStore;
+
+
+    //TODO ?? Nastavení v menu/tools -> YAML nebo XML https://learn.microsoft.com/en-us/troubleshoot/developer/visualstudio/csharp/language-compilers/store-custom-information-config-file
+    public IconApp() : base("File Commander")
     {
-        //Defaults
-        SetDefaultSize(720, 512);
+        SetDefaultSize(1280, 720);
+        Maximize();
         SetPosition(WindowPosition.Center);
         DeleteEvent += delegate { Application.Quit(); };
-        
-        
-        
-        MenuBar mb = new MenuBar();
-        Menu filemenu = new Menu();
-        MenuItem file = new MenuItem("File");
-        file.Submenu = filemenu;
-       
-        AccelGroup agr = new AccelGroup();
-        AddAccelGroup(agr);
 
-        ImageMenuItem newi = new ImageMenuItem(Stock.New, agr);
-        newi.AddAccelerator("activate", agr, new AccelKey(
-            Gdk.Key.n, Gdk.ModifierType.ControlMask, AccelFlags.Visible));
-        filemenu.Append(newi);
 
-        ImageMenuItem open = new ImageMenuItem(Stock.Open, agr);
-        open.AddAccelerator("activate", agr, new AccelKey(
-            Gdk.Key.n, Gdk.ModifierType.ControlMask, AccelFlags.Visible));
-        filemenu.Append(open);
+        //Vytvoření kontejneru vnitřního obsahu okna
+        VBox windowVerticalBox = new VBox(false, 0);
+        Add(windowVerticalBox);
 
-        SeparatorMenuItem menusep = new SeparatorMenuItem();
-        filemenu.Append(menusep);
+        #region Toolbar
 
-        ImageMenuItem exit = new ImageMenuItem(Stock.Quit, agr);
-        exit.AddAccelerator("activate", agr, new AccelKey(
-            Gdk.Key.q, Gdk.ModifierType.ControlMask, AccelFlags.Visible));
+        //Vytvoření hlavní nástrojové lišty
+        var toolbar = new Toolbar();
+        toolbar.ToolbarStyle = ToolbarStyle.Both;
 
-        exit.Activated += OnActivated;
-        filemenu.Append(exit);
+        var toolRefreshButton = new ToolButton(Stock.Refresh);
+        toolbar.Insert(toolRefreshButton, 0);
+        toolRefreshButton.Clicked += OnRefreshClicked!;
 
-        mb.Append(file);
+        var toolNewButton = new ToolButton(Stock.New);
+        toolbar.Insert(toolNewButton, 6);
+        toolNewButton.Clicked += OnNewClicked!;
 
-        VBox menuvbox = new VBox(false, 2);
-        menuvbox.PackStart(mb, false, false, 0);
-        menuvbox.PackStart(new Label(), false, false, 0);
-        
-        Add(menuvbox);
-        
-        //Toolbar
-        Toolbar toolbar = new Toolbar();
-        toolbar.ToolbarStyle = ToolbarStyle.Icons;
+        var toolCopyButton = new ToolButton(Stock.Copy);
+        toolbar.Insert(toolCopyButton, 7);
+        toolCopyButton.Clicked += OnCopyClicked!;
 
-        ToolButton newtb = new ToolButton(Stock.New);
-        ToolButton opentb = new ToolButton(Stock.Open);
-        ToolButton savetb = new ToolButton(Stock.Save);
-        SeparatorToolItem toolsep = new SeparatorToolItem();
-        ToolButton quittb = new ToolButton(Stock.Quit);
+        var toolMoveButton = new ToolButton(Stock.Dnd); //Icon TBA
+        toolbar.Insert(toolMoveButton, 8);
+        toolMoveButton.Clicked += OnMoveClicked!;
 
-        toolbar.Insert(newtb, 0);
-        toolbar.Insert(opentb, 1);
-        toolbar.Insert(savetb, 2);
-        toolbar.Insert(toolsep, 3);
-        toolbar.Insert(quittb, 4);
+        var toolRenameButton = new ToolButton(Stock.Network); //Icon TBA
+        toolbar.Insert(toolRenameButton, 9);
+        toolRenameButton.Clicked += OnRenameClicked!;
 
-        quittb.Clicked += OnClicked;
+        var toolDeleteButton = new ToolButton(Stock.Delete);
+        toolbar.Insert(toolDeleteButton, 10);
+        toolDeleteButton.Clicked += OnDeleteClicked!;
+
+        var toolExtractButton = new ToolButton(Stock.Execute); //Icon TBA
+        toolbar.Insert(toolExtractButton, 12);
+        toolExtractButton.Clicked += OnExtractClicked!;
+
+        var toolCompressButton = new ToolButton(Stock.Execute); //Icon TBA
+        toolbar.Insert(toolCompressButton, 12);
+        toolCompressButton.Clicked += OnCompressClicked!;
+
+        toolbar.Insert(new SeparatorToolItem(), 5);
+        toolbar.Insert(new SeparatorToolItem(), 11);
+
+        windowVerticalBox.PackStart(toolbar, false, false, 0);
+
+        #endregion
+
+        //Vytvoření lišty správy adresářových panelů
+        HBox twinPanelToolbox = new HBox();
+        //TODO list dostupných disků - https://learn.microsoft.com/en-us/dotnet/api/system.io.driveinfo.getdrives?redirectedfrom=MSDN&view=net-7.0#System_IO_DriveInfo_GetDrives
+
+
+        #region LeftTwinBar
+
+        HBox leftTwinToolbox = new HBox();
+
+        //LEVÁ LIŠTA
+        var leftToolbar = new Toolbar();
+        leftToolbar.ToolbarStyle = ToolbarStyle.Both;
+
+        var leftHomeButton = new ToolButton(Stock.Home);
+        leftToolbar.Insert(leftHomeButton, 0);
+        leftHomeButton.Clicked += delegate(object? sender, EventArgs args)
+        {
+            LeftRoot = OnHomeClicked(sender, args, LeftRoot, LeftStore);
+        };
+
+        var leftUpButton = new ToolButton(Stock.GoUp);
+        leftToolbar.Insert(leftUpButton, 1);
+        leftUpButton.Clicked += delegate { LeftRoot = OnUpClicked(LeftRoot, LeftStore); };
+
+        leftToolbar.Insert(new SeparatorToolItem(), 2);
+
+        var leftToolBackButton = new ToolButton(Stock.GoBack);
+        leftToolbar.Insert(leftToolBackButton, 3);
+        leftToolBackButton.Clicked += OnBackClicked!;
+
+        var leftToolForwardButton = new ToolButton(Stock.GoForward);
+        leftToolbar.Insert(leftToolForwardButton, 4);
+        leftToolForwardButton.Clicked += OnForwardClicked!;
+
+        var leftToolUndoButton = new ToolButton(Stock.Undo);
+        leftToolbar.Insert(leftToolUndoButton, 5);
+        leftToolUndoButton.Clicked += OnUndoClicked!;
+
+        var leftToolRedoButton = new ToolButton(Stock.Redo);
+        leftToolbar.Insert(leftToolRedoButton, 5);
+        leftToolRedoButton.Clicked += OnRedoClicked!;
+
+        //LEVÝ SEZNAM DISKŮ
+        /* HBox leftDisksBar = new HBox();
          
-        VBox toolvbox = new VBox(false, 2);
-        toolvbox.PackStart(toolbar, false, false, 0);
+         var availableDrives = DriveInfo.GetDrives();
+         List<DriveInfo> drives = new List<DriveInfo>();
+ 
+         var pos = 0; //
+         foreach (var drive in availableDrives)
+         {
+             if (drive.IsReady)
+             {
+                 Console.WriteLine(drive.VolumeLabel);
+                 var necosl = new ToolButton(drive.VolumeLabel);
+                 leftToolbar.Insert(necosl, 6);
+                 necosl.Clicked += delegate { Console.WriteLine("clicked necosl"); };
+                 //var driveButton = new Button(drive.VolumeLabel);
+                 //leftDisksBar.Add(driveButton);
+             }
+         }
+ 
+      */
+        twinPanelToolbox.PackStart(leftToolbar, true, true, 0);
 
-        Add(toolvbox);
-        
+        #endregion
+
+        #region RightTwinBar
+
+        //PRAVÁ LIŠTA
+        var rightPanelBar = new Toolbar();
+        rightPanelBar.ToolbarStyle = ToolbarStyle.Both;
+
+        var rightHomeButton = new ToolButton(Stock.Home);
+        rightPanelBar.Insert(rightHomeButton, 0);
+        rightHomeButton.Clicked += delegate(object? sender, EventArgs args)
+        {
+            RightRoot = OnHomeClicked(sender, args, RightRoot, RightStore);
+        };
+
+        var rightUpButton = new ToolButton(Stock.GoUp);
+        rightPanelBar.Insert(rightUpButton, 1);
+        rightUpButton.Clicked += delegate { RightRoot = OnUpClicked(RightRoot, RightStore); };
+
+        rightPanelBar.Insert(new SeparatorToolItem(), 2);
+
+        var rightToolBackButton = new ToolButton(Stock.GoBack);
+        rightPanelBar.Insert(rightToolBackButton, 3);
+        rightToolBackButton.Clicked += OnBackClicked!;
+
+        var rightToolForwardButton = new ToolButton(Stock.GoForward);
+        rightPanelBar.Insert(rightToolForwardButton, 4);
+        rightToolForwardButton.Clicked += OnForwardClicked!;
+
+        var rightToolUndoButton = new ToolButton(Stock.Undo);
+        rightPanelBar.Insert(rightToolUndoButton, 5);
+        rightToolUndoButton.Clicked += OnUndoClicked!;
+
+
+        var rightToolRedoButton = new ToolButton(Stock.Redo);
+        rightPanelBar.Insert(rightToolRedoButton, 6);
+        rightToolRedoButton.Clicked += OnRedoClicked!;
+
+
+        twinPanelToolbox.PackStart(rightPanelBar, true, true, 0);
+
+        #endregion
+
+        windowVerticalBox.PackStart(twinPanelToolbox, false, true, 0);
+
+        #region LeftIconView
+
+        //Levý panel
+        ScrolledWindow leftScrolledWindow = new ScrolledWindow();
+        leftScrolledWindow.ShadowType = ShadowType.EtchedIn;
+        leftScrolledWindow.SetPolicy(PolicyType.Automatic, PolicyType.Automatic);
+        LeftStore = CreateStore();
+        FillStore(LeftStore, LeftRoot);
+
+        IconView leftIconView = new IconView(LeftStore);
+        leftIconView.SelectionMode = SelectionMode.Multiple;
+        leftIconView.TextColumn = ColDisplayName;
+        leftIconView.PixbufColumn = ColPixbuf;
+        leftIconView.ItemActivated += delegate(object _, ItemActivatedArgs args)
+        {
+            LeftRoot = OnItemActivated(args, LeftRoot, LeftStore);
+        };
+
+        leftScrolledWindow.Add(leftIconView);
+        //leftIconView.GrabFocus();
+
+        #endregion
+
+        #region RightIconView
+
+        //Pravý panel
+        ScrolledWindow rightScrolledWindow = new ScrolledWindow();
+        rightScrolledWindow.ShadowType = ShadowType.EtchedIn;
+        rightScrolledWindow.SetPolicy(PolicyType.Automatic, PolicyType.Automatic);
+        RightStore = CreateStore();
+        FillStore(RightStore, RightRoot);
+
+        IconView rightIconView = new IconView(RightStore);
+        rightIconView.SelectionMode = SelectionMode.Multiple;
+        rightIconView.TextColumn = ColDisplayName;
+        rightIconView.PixbufColumn = ColPixbuf;
+        rightIconView.ItemActivated += delegate(object _, ItemActivatedArgs args)
+        {
+            RightRoot = OnItemActivated(args, RightRoot, RightStore);
+        };
+
+        rightScrolledWindow.Add(rightIconView);
+        //rightIconView.GrabFocus();
+
+        #endregion
+
+        HBox twinPanelsBox = new HBox(false, 0);
+        twinPanelsBox.PackStart(leftScrolledWindow, true, true, 0);
+        twinPanelsBox.PackStart(rightScrolledWindow, true, true, 0);
+
+        windowVerticalBox.PackStart(twinPanelsBox, true, true, 0);
+
 
         ShowAll();
     }
 
-    void OnActivated(object sender, EventArgs args)
+    public static Gdk.Pixbuf GetIcon(string name)
     {
-        Application.Quit();
+        return IconTheme.Default.LoadIcon(name, 48, 0);
     }
-    
-    void OnClicked(object sender, EventArgs args)
+
+    private static ListStore CreateStore()
     {
-        Application.Quit();
+        ListStore store = new ListStore(typeof(string),
+            typeof(string), typeof(Gdk.Pixbuf), typeof(bool));
+
+        store.SetSortColumnId(ColDisplayName, SortType.Ascending);
+
+        return store;
+    }
+
+    public static void FillStore(ListStore store, DirectoryInfo root)
+    {
+        store.Clear();
+
+        if (!root.Exists)
+        {
+            return;
+        }
+
+        foreach (DirectoryInfo di in root.GetDirectories())
+        {
+            if (!di.Name.StartsWith("."))
+                store.AppendValues(di.FullName, di.Name, DirIcon, true);
+        }
+
+        foreach (FileInfo file in root.GetFiles())
+        {
+            if (!file.Name.StartsWith("."))
+                store.AppendValues(file.FullName, file.Name, FileIcon, false);
+        }
+    }
+
+    private static DirectoryInfo OnItemActivated(ItemActivatedArgs a, DirectoryInfo root, ListStore store)
+    {
+        TreeIter iter;
+        store.GetIter(out iter, a.Path);
+        var path = (string) store.GetValue(iter, ColPath);
+        var isDir = (bool) store.GetValue(iter, ColIsDirectory);
+
+        if (!isDir)
+            return root;
+
+        root = new DirectoryInfo(path);
+        FillStore(store, root);
+        return root;
     }
 }
